Fun with C strings

+++ src/reference.hpp
@@ -121,7 +121,7 @@ private:
 template <typename T>
 Eigen::MatrixXf kmer2mat(const std::vector<T> &kmers) {
   Eigen::MatrixXf kmer_mat = Eigen::MatrixXf::Ones(kmers.size(), 2);
-  for (int i = 0; i < kmers.size(); ++i) {
+  for (size_t i = 0; i < kmers.size(); ++i) {
     kmer_mat(i, 1) = kmers[i];
   }
   return kmer_mat;
